% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicles-data.R
\docType{data}
\name{vehicles}
\alias{vehicles}
\title{Vehicles dataset}
\format{A list with two elements:
\describe{
\item{x}{The 846 x 18 object-attribute matrix.}
\item{y}{A 846-vector containing the class labels.}
}}
\usage{
data(vehicles)
}
\description{
This dataset was collected from silhouettes by the HIPS (Hierarchical Image
Processing System) extension BINATTS  Four model vehicles were used for the
experiment: bus, Chevrolet van,  Saab 9000 and Opel Manta. The data were used to
distinguish 3D objects within a 2-D silhouette of the objects. There are 846
instances and 18 numeric attributes. The first 564 objects are training data,
the rest are test data. This version of dataset was used by Zouhal and Denoeux (1998).
}
\examples{
data(vehicles)
table(vehicles$y)
}
\references{
P. M.  Murphy and D. W. Aha.  UCI Reposition of machine learning databases.
[Machine readable data repository]. University of California, Departement of
Information and Computer Science, Irvine, CA.

L. M. Zouhal and T. Denoeux. An evidence-theoretic k-NN rule with parameter
optimization. IEEE Transactions on Systems, Man and Cybernetics Part C,
28(2):263--271,1998.
}
\keyword{datasets}

